Cleanup: move to extension package, renaming
