+++ otel-extension/src/main/java/akka/actor/impl/LocalActorRefProviderAdvice.java
@@ -2,7 +2,7 @@ package akka.actor.impl;
 
 import akka.actor.ActorRef;
 import akka.actor.ActorSystem;
-import io.scalac.mesmer.otelextension.instrumentations.akka.actor.ActorLifecycleEvent;
+import io.scalac.mesmer.otelextension.instrumentations.akka.actor.extension.ActorLifecycleEvents;
 import net.bytebuddy.asm.Advice;
 
 public class LocalActorRefProviderAdvice {
@@ -10,6 +10,6 @@ public class LocalActorRefProviderAdvice {
   @Advice.OnMethodExit
   public static void actorOfExit(
       @Advice.Argument(0) ActorSystem system, @Advice.Return ActorRef ref) {
-    system.eventStream().publish(new ActorLifecycleEvent.ActorCreated(ref));
+    system.eventStream().publish(new ActorLifecycleEvents.ActorCreated(ref));
   }
 }
