+++ otel-extension/src/main/scala/io/scalac/mesmer/otelextension/instrumentations/akka/actor/extension/ActorLifecycleEvents.scala
@@ -1,0 +1,10 @@
+package io.scalac.mesmer.otelextension.instrumentations.akka.actor.extension
+
+import akka.actor.ActorRef
+
+trait ActorLifecycleEvents
+
+object ActorLifecycleEvents {
+  final case class ActorCreated(ref: ActorRef)    extends ActorLifecycleEvents
+  final case class ActorTerminated(ref: ActorRef) extends ActorLifecycleEvents
+}
